+++ Api Rest/municipalidad/src/main/java/org/una/municipalidad/dto/Contribuyentes_Locales_MercadoDTO.java
@@ -11,7 +11,7 @@ import org.una.municipalidad.entities.Locales_Mercado;
 @Builder
 public class Contribuyentes_Locales_MercadoDTO {
     private Long id;
-    private Long porcentajeLocales;
+    private Double porcentajeLocales;
     private ContribuyentesDTO contribuyente;
     private LocalesMercadoDTO localesmercado;
 }
